pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "bcrypt>=4.3.0",
    "numpy>=2.3.3",
    "openai>=1.107.2",
    "openpyxl>=3.1.5",
    "pandas>=2.3.2",
    "pdfplumber>=0.11.7",
    "plotly>=6.3.0",
    "psycopg2-binary>=2.9.10",
    "pypdf2>=3.0.1",
    "python-multipart>=0.0.20",
    "streamlit>=1.49.1",
]
